requirements.txt
streamlit
scikit-learn
PyPDF2
python-docx
pandas
numpy
matplotlib
plotly
transformers
sentence-transformers
torch


keras<3.0
tf-keras
sentence-transformers
transformers
streamlit
pandas
plotly
numpy